requirements.txt
fastapi[standard]
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pydantic
